* commit 'eb485256e99e97653773024951568aadcc60f6a4':
  Camera2 Tests: Fix stream id types

+++ tests/camera2/CameraFrameTests.cpp
@@ -90,7 +90,7 @@ TEST_P(CameraFrameTest, GetFrame) {
     ASSERT_EQ(OK, mDevice->createDefaultRequest(CAMERA2_TEMPLATE_PREVIEW,
                                                 &previewRequest));
     {
-        Vector<uint8_t> outputStreamIds;
+        Vector<int32_t> outputStreamIds;
         outputStreamIds.push(mStreamId);
         ASSERT_EQ(OK, previewRequest.update(ANDROID_REQUEST_OUTPUT_STREAMS,
                                             outputStreamIds));
